fixed: samba can't handle files > 2 gb correctly
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@522 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/FileItem.h
@@ -22,7 +22,7 @@ public:
   bool          m_bIsShareOrDrive;		// is this a root share/drive
   int			m_iDriveType;			// if it is a root share/drive which type. Types see CShare
   SYSTEMTIME    m_stTime;							// file creation date & time
-  DWORD         m_dwSize;							// file size (0 for folders)
+  __int64       m_dwSize;							// file size (0 for folders)
 	CMusicInfoTag m_musicInfoTag;
 };
